decouple description writing
Here's the dilemma I've been having: every time the `run_test()`
function is called, it `fork()`s into a child process and goes
into `_run_test()`, where a `TestResult` `struct` is written and
written back via `pipe()` `FILE` descriptor.

The `run_test()` function is capable of telling whether its
`_run_test()` child crashed or not, but the issue is that whenever
the `_run_test()` child crashes, it does not end up writing a
`TestResult` struct. This is an issue because a `TestResult`
consists of the `result`, `comment`, and to my chagrin, the
`description` as well.

The `description` of a test exists regardless of the `result` or
`comment` (as the `comment` exists only to explain the `result`,
which means it varies between test results). This meant that the
description was very coupled within the testing process, and that
whenever the test crashed, it would not be able to send it.

Now, I'm very stubborn and don't want to refactor my code, so here
is what I did:

1. I turned the `pipe()` `FILE` descriptors into a global variable
in order for both `seer.c` and `test_cases.c` to access it for
`write()` purposes. I did this by moving `seer.c`'s `fd` array
to outside the `run_test()` function to the file's scope and
putting an `extern` linkage definition in `seer.h`.

2. From now on, every time a test case is executed in the
instructor's file, rather than write to the `TestResult` `struct`
directly, the maintainer will have to call `description()`, which
writes to the `FILE` descriptors created by the `pipe()`. This
is easier than the previous implementation (which was to call
`strcpy` onto the intermediate TestResult). Only then will the
`_run_test()` function write the actual struct. This effectively
separates `result` and `comment` from `description` until
`run_test()` inevitably reads both the `description` and
intermediate `TestResult` back and combines the two. Not sure if
genius or braindead.

3. As stated parent process that invigilates the test will now
read in the `description` from the pipe, and then the actual
`TestResult` `struct`.

4. The parent process combines the two into an actual `TestResult`
`struct` to be returned from `run_test()`.

Because this "intermediate splicing" only occurs between the
`run_test()` and `_run_test()` functions, there shouldn't be an
issue with anyone calling `run_test()` as `_run_test()` is an
internal function hidden from the user.

After these weird Frankenstein-esque operations are done, the
`run_test()` function effectively returns a fully-dressed
`TestResult` `struct` without any fuss.

_If you've made it this far in my commit message, congrats._

Because this process requires two `read()`s and two `write()`s,
I  had to be extra careful about it. Which is why I made the
`description()` function to ensure that exactly `MAX_STR` bytes
were written when the description was sent (in order to ensure
alignment with the `TestResult` `struct` being written). This
function will write an error message in the event that a message
longer than `MAX_STR` is sent.

Along with that, I added extra checks to both the parent and child
`fork()` code in `run_test()`. This is where you'll really want to
check the return value(s) of such low-level functions (`read()`
and `write()`). These system calls have absolutely no chill and
will write whatever you tell them to.

The parent function will always assert that it reads `MAX_STR`
bytes first for the description, and then `sizeof(TestResult)`
bytes for the actual test result `struct`. If the first `read()`
of `MAX_STR` `description` bytes fails, then something really
bad has happened, as that means even the `TestStruct` wasn't
written. At that point I `abort()` the entire program because
clearly I did something bad in development. Please contact me
if you get an `abort()` message.

Now, to the topic of alignment, if the `description` is `read()`
properly but the `TestResult` isn't, that means the `description`
wasn't written. How do I know this? Because we expect `MAX_STR`
bytes to be written first, and then `sizeof(TestResult)`. If the
`description` isn't written, but the `TestResult` is, then the
first `read()` from the parent is actually an attempt to `read()`
the `TestStruct` bytes into the `description`, which is basically
garbage data. In the event that the `TestStruct` could not be
`read()`, I put an error message reminding the invigilator to make
sure a `description()` is added.

Time will tell if this is a good idea.